Moved functions to init
Moved a lot of code from main into init so that the app is properly initilized
